[#192] MODIFY: ConfirmRemoveProfileView view, controller 분리
